+++ common/database_utils.h
@@ -16,11 +16,11 @@ int logout_user(std::string user_id);
 int unregister_user(std::string user_id);
 time_t get_auction_start_and_end_fulltime(std::string AID, char mode);
 int create_auction_dirs(std::string AID, std::string UID);
-void copy_image(std::string AID, std::string file_name, int file_size, char* image_buffer, char* image);
-void create_auction_start_file(std::string AID, std::string UID, std::string asset_name, std::string file_name, int start_value, int timeactive);
+void copy_image(std::string AID, std::string file_name, long file_size, char* image_buffer, char* image);
+void create_auction_start_file(std::string AID, std::string UID, std::string asset_name, std::string file_name, long start_value, long timeactive);
 int create_auction_end_file(std::string AID);
 std::list <std::string> get_hosted_auctions(std::string UID);
-int get_highest_bid(std::string AID, int new_bid);
-int create_bid_files(std::string UID, std::string AID, int value, std::string value_str);
+int get_highest_bid(std::string AID, long new_bid);
+int create_bid_files(std::string UID, std::string AID, long value, std::string value_str);
 
 #endif
